crypto: move signify to common dir
